Add test for calling Nif.create_root_span/1 in Tracer.create_span/1
Although a reference is asserted in the tests for Tracer.create_span/1
already, this patch adds a regression test to make sure the reference
comes from the Nif.

It adds Appsignal.WrappedNif, a test module that wraps Appsignal.Nif. It
calls to the orignal module like normal, but registers the call to allow
assertions on later. Currently, only WrappedNif.create_root_span/1 is
implemented. Other functions will be implemented as needed.